feat: use an AWS-provided role for SSM administration
We don't have an easy way to figure out what permissions the AWS provider needs in order to manipulate the resources we're decalring in all CRUD situations. Since we don't really need to tighten perms down as far as we had it, we'll use this wider role. We can tighten it down again in future if necessary.