new: delete solr6 app/eks cluster
To save on costs, we are removing obsolete code.  We no longer need solr 6 and we are not using EKS anymore.  The solrcloud app still needs a k8s cluster to be able to start up, so we need to provision the bare minimum; but overall, this should save on costs and give us less to maintain